Refactored Pie chart to separate component
